% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr.R
\name{clr}
\alias{clr}
\title{Clusterwise Linear Regression}
\usage{
clr(X, ...)
}
\arguments{
\item{X}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Clusterwise Linear Regression
}
\seealso{
\code{\link{clr.default}}.
}
