% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{clr_lloyd_cpp}
\alias{clr_lloyd_cpp}
\title{Clusterwise Linear Regression}
\usage{
clr_lloyd_cpp(U, X, y, m, itermax)
}
\arguments{
\item{U}{random membership matrix.}

\item{X}{data matrix of dimension \code{N} by \code{P}.}

\item{y}{response vector of length \code{N}.}

\item{m}{fuzzier exponent. Gives a hard k-partition if \code{m = 1.0} and
a fuzzy k-partition when \code{m > 1.0}.}

\item{itermax}{the maximum number of iterations allowed in Lloyd algorithm.}
}
\value{
returns an object of class \code{WCLR.clr}.
}
\description{
This is a bare-bones function and must NOT be used by an end-user.
}
\seealso{
\code{\link{clr.default}} for a user-friendly version of this method.
}
