% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{fuzzy_kpartition}
\alias{fuzzy_kpartition}
\title{Fuzzy K-partition}
\usage{
fuzzy_kpartition(N, K)
}
\arguments{
\item{N}{number of observations}

\item{K}{number of clusters}
}
\description{
Makes a random fuzzy k-partition membership matrix of dimension
\code{N} by \code{K}.
}
\seealso{
\code{\link{hard_kpartition}} for a hard (crisp) k-partition.
}
