% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{hard_kpartition}
\alias{hard_kpartition}
\title{Hard K-partition}
\usage{
hard_kpartition(N, K)
}
\arguments{
\item{N}{number of observations}

\item{K}{number of clusters}
}
\description{
Makes a random hard (crisp) k-partition membership matrix of dimension
\code{N} by \code{K}.
}
\seealso{
\code{\link{fuzzy_kpartition}} for a fuzzy k-partition.
}
