% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{kmeans.default}
\alias{kmeans.default}
\title{K-means Clustering + Linear Regression}
\usage{
\method{kmeans}{default}(
  X,
  y,
  K,
  m = 1,
  nstart = 1L,
  iter.max = 100L,
  algorithm = c("Lloyd"),
  ...
)
}
\arguments{
\item{X}{data matrix of dimension \code{N} by \code{P}.}

\item{y}{response vector of length \code{N}.}

\item{K}{number of clusters.}

\item{m}{fuzzier exponent. Gives a hard k-partition if \code{m = 1.0} and
a fuzzy k-partition when \code{m > 1.0}.}

\item{nstart}{number of random runs.}

\item{iter.max}{the maximum number of iterations allowed in Lloyd algorithm.}

\item{algorithm}{optimization algorithm. One of: \code{"Lloyd"}.}

\item{...}{not used.}
}
\value{
returns an object of class \code{wclr.kmeans}.
}
\description{
K-means Clustering + Linear Regression
}
