% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{kmeans}
\alias{kmeans}
\title{K-means Clustering + Linear Regression}
\usage{
kmeans(X, ...)
}
\arguments{
\item{X}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
K-means Clustering + Linear Regression
}
\seealso{
\code{\link{kmeans.default}}.
}
