% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kmeans_lloyd_cpp}
\alias{kmeans_lloyd_cpp}
\title{K-means Linear Regression}
\usage{
kmeans_lloyd_cpp(U, X, y, m, itermax, trace)
}
\arguments{
\item{U}{random membership matrix.}

\item{X}{data matrix of dimension \code{N} by \code{P}.}

\item{y}{response vector of length \code{N}.}

\item{m}{fuzzier exponent. Gives a hard k-partition if \code{m = 1.0} and
a fuzzy k-partition when \code{m > 1.0}.}

\item{itermax}{the maximum number of iterations allowed in Lloyd algorithm.}

\item{trace}{logical value. If \code{TRUE}, produce a trace information of
the progress of the algorithm.}
}
\value{
returns an object of class \code{wclr.kmeans}.
}
\description{
This is a bare-bones function and must NOT be used by an end-user.
}
\seealso{
\code{\link{kmeans.default}} for a user-friendly version of this method.
}
