% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kplane.R
\name{kplane}
\alias{kplane}
\title{K-plane Regression}
\usage{
kplane(X, ...)
}
\arguments{
\item{X}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
K-plane Regression
}
\seealso{
\code{\link{kplane.default}}.
}
