% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{predict.wclr}
\alias{predict.wclr}
\title{Predict}
\usage{
\method{predict}{wclr}(object, newdata, ...)
}
\arguments{
\item{object}{object of class inheriting from "wclr".}

\item{newdata}{a data matrix in which to look for variables with
which to predict.}

\item{...}{not used.}
}
\value{
produces a vector of predictions.
}
\description{
Predicted values based on \code{\link{wclr.default}} model object.
}
