% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wclr.R
\name{predict.WCLR.wclr}
\alias{predict.WCLR.wclr}
\title{Predict}
\usage{
\method{predict}{WCLR.wclr}(object, newdata, ...)
}
\arguments{
\item{object}{object of class inheriting from "WCLR.wclr".}

\item{newdata}{a data matrix in which to look for variables with
which to predict.}

\item{...}{not used.}
}
\value{
produces a vector of predictions.
}
\description{
Predicted values based on \code{\link{swclr.default}} model object.
}
