% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predict_euclidean_cpp}
\alias{predict_euclidean_cpp}
\title{Predict}
\usage{
predict_euclidean_cpp(X, C, R, m)
}
\arguments{
\item{X}{newdata matrix}

\item{C}{centers matrix}

\item{R}{coefficients matrix}

\item{m}{fuzzier exponent}
}
\value{
a vector of predictions.
}
\description{
Predict values by nearest Euclidean distance.
}
