% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predict_quadratic_cpp}
\alias{predict_quadratic_cpp}
\title{Predict}
\usage{
predict_quadratic_cpp(X, C, W, R, m)
}
\arguments{
\item{X}{newdata matrix}

\item{C}{centers matrix}

\item{W}{weights cube}

\item{R}{coefficients matrix}

\item{m}{fuzzier exponent}
}
\value{
a vector of predictions.
}
\description{
Predict values by nearest quadratic distance.
}
