% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swclr.R
\name{swclr.default}
\alias{swclr.default}
\title{Self-balanced Weighted Clusterwise Linear Regression}
\usage{
\method{swclr}{default}(
  X,
  y,
  K,
  balance,
  wnorm,
  m = 1,
  nstart = 1L,
  iter.max = 100L,
  algorithm = c("Lloyd"),
  ...
)
}
\arguments{
\item{X}{data matrix of dimension \code{N} by \code{P}.}

\item{y}{response vector of length \code{N}.}

\item{K}{number of clusters.}

\item{balance}{self-balance type. One of: \code{"pg"}, \code{"pl"}.}

\item{wnorm}{wnorm type. One of: \code{"epg"}, \code{"epl"},
\code{"qpg"}, \code{"qpl"}.}

\item{m}{fuzzier exponent. Gives a hard k-partition if \code{m = 1.0} and
a fuzzy k-partition when \code{m > 1.0}.}

\item{nstart}{number of random runs.}

\item{iter.max}{the maximum number of iterations allowed in Lloyd algorithm.}

\item{algorithm}{optimization algorithm. One of: \code{"Lloyd"}.}

\item{...}{not used.}
}
\value{
returns an object of class \code{wclr.swclr}.
}
\description{
Self-balanced Weighted Clusterwise Linear Regression
}
