% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swclr.R
\name{swclr}
\alias{swclr}
\title{Self-balanced Weighted Clusterwise Linear Regression}
\usage{
swclr(X, ...)
}
\arguments{
\item{X}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Self-balanced Weighted Clusterwise Linear Regression
}
\seealso{
\code{\link{swclr.default}}.
}
