% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{swclr_lloyd_cpp}
\alias{swclr_lloyd_cpp}
\title{Self-balanced Weighted Clusterwise Linear Regression}
\usage{
swclr_lloyd_cpp(U, X, y, m, wnorm, balance, itermax, trace)
}
\arguments{
\item{U}{random membership matrix.}

\item{X}{data matrix of dimension \code{N} by \code{P}.}

\item{y}{response vector of length \code{N}.}

\item{m}{fuzzier exponent. Gives a hard k-partition if \code{m = 1.0} and
a fuzzy k-partition when \code{m > 1.0}.}

\item{wnorm}{wnorm type. One of: \code{"epg"}, \code{"epl"},
\code{"qpg"}, \code{"qpl"}.}

\item{balance}{self-balance type. One of: \code{"pg"}, \code{"pl"}.}

\item{itermax}{the maximum number of iterations allowed in Lloyd algorithm.}

\item{trace}{logical value. If \code{TRUE}, produce a trace information of
the progress of the algorithm.}
}
\value{
returns an object of class \code{WCLR.swclr}.
}
\description{
This is a bare-bones function and must NOT be used by an end-user.
}
\seealso{
\code{\link{swclr.default}} for a user-friendly version of this method.
}
