% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wclr-package.R
\docType{package}
\name{WCLR-package}
\alias{WCLR}
\alias{WCLR-package}
\title{WCLR: Weighted Clusterwise Linear Regression}
\description{
Provides a weighted clusterwise linear regression method to data fitting. It also implements many related clusterwise linear regression methods.
}
\author{
\strong{Maintainer}: Ricardo Azevedo \email{ricardo.azevedo@univasf.edu.br} (\href{https://orcid.org/0000-0001-6703-2427}{ORCID})

}
\keyword{internal}
