% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wclr.R
\name{wclr}
\alias{wclr}
\title{Weighted Clusterwise Linear Regression}
\usage{
wclr(X, ...)
}
\arguments{
\item{X}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Weighted Clusterwise Linear Regression
}
\seealso{
\code{\link{wclr.default}}.
}
