% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{predict.WCLR}
\alias{predict.WCLR}
\title{Predict values}
\usage{
\method{predict}{WCLR}(object, newdata, ...)
}
\arguments{
\item{object}{object of class inheriting from "WCLR".}

\item{newdata}{a data matrix in which to look for variables with
which to predict.}

\item{...}{not used.}
}
\value{
produces a vector of predictions.
}
\description{
Predict values
}
