% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swclr.R
\name{predict.WCLR.swclr}
\alias{predict.WCLR.swclr}
\title{Predict}
\usage{
\method{predict}{WCLR.swclr}(object, newdata, ...)
}
\arguments{
\item{object}{object of class inheriting from "WCLR.swclr".}

\item{newdata}{a data matrix in which to look for variables with
which to predict.}

\item{...}{not used.}
}
\value{
produces a vector of predictions.
}
\description{
Predicted values based on \code{\link{swclr.default}} model object.
}
